Update projects.html
been stumped all week on which style work page to use.  I like both the week 5 video one and the one we did previously.  I like the style we did previously a bit better. I'm going to hand it in in that style.  However I did add an H2 like the one in the vid--more seo friendly